app/requirements.txt
fastapi==0.112.0
uvicorn==0.31.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.8
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.2